refactor: 타입 재사용
