+++ client/src/types/auth.ts
@@ -10,3 +10,22 @@ export interface IAuth {
   userId: IUserId;
   userLoading: boolean;
 }
+
+export interface IAuthState {
+  id: string;
+  password: string;
+}
+
+export interface IReceiveServer {
+  accessToken: string;
+  userId: string;
+}
+
+export interface ICheckUser {
+  newAccessToken?: string;
+  userId: string;
+}
+
+export interface IGithubCode {
+  code: string;
+}
